Move all_questions from controller to quiz service
